refactor: change utils path
